// src/components/Title/types.d.ts
export interface IContentProps {
  small?: boolean;
  medium?: boolean;
  large?: boolean;
  primary?: boolean;
  weight?: boolean;
  title: string;
  uppercase?: boolean;
  decoration?: boolean;
  red?: boolean;
  blue?: boolean;
}

export interface IPrimaryColorProps {
  primary?: boolean;
  weight?: boolean;
  uppercase?: boolean;
  red?: boolean;
  blue?: boolean;
}